webapp/pyproject.toml
[tool.poetry]
name = "cleansecpy-web"
version = "0.1.0"
description = ""
authors = ["Luidgi Gromat <gromat.luidgi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
whitenoise = "^6.1.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
